requirements.txt
openai
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
playwright
trafilatura
chromadb
pypdf
pyyaml